+++ src/nbody.cpp
@@ -1,13 +1,17 @@
-#include <Rcpp.h>
+#include <cpp11/doubles.hpp>
+#include <cpp11/matrix.hpp>
 #include "vector.h"
 #include "quadTree.h"
 
-using namespace Rcpp;
-
-//[[Rcpp::export]]
-NumericMatrix nbody(NumericMatrix pos, NumericVector strength, double theta, double min_dist, double max_dist, double alpha) {
+[[cpp11::register]]
+cpp11::writable::doubles_matrix<> nbody_c(cpp11::doubles_matrix<> pos,
+                                          cpp11::doubles strength,
+                                          double theta,
+                                          double min_dist,
+                                          double max_dist,
+                                          double alpha) {
   size_t i;
-  NumericMatrix res(pos.nrow(), pos.ncol());
+  cpp11::writable::doubles_matrix<> res(pos.nrow(), pos.ncol());
   QuadTree<2> tree(theta, min_dist, max_dist, alpha);
   std::deque<Body<2> *> bodies;
   for (i = 0; i < pos.nrow(); ++i) {