+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Fix a bug in trap_force with incorrect indexing in certain circumstances
 * Added `replace_particles()`
+* Moved to cpp11
 
 # particles 0.2.2